+++ main.go
@@ -53,6 +53,7 @@ func main() {
 	subcommands.Register(&cmd.Search{}, "")
 	subcommands.Register(&cmd.SetDone{}, "")
 	subcommands.Register(&cmd.Stats{}, "")
+	subcommands.Register(&cmd.UpdateTitles{}, "")
 	subcommands.Register(&cmd.Unfinished{}, "")
 	subcommands.Register(&cmd.Unregister{}, "")
 	subcommands.Register(&cmd.Watch{}, "")